Update existing events to use the new BaseEvent
