+++ pizzaserver-server/src/main/java/io/github/willqi/pizzaserver/server/event/type/player/PreLoginEvent.java
@@ -1,8 +1,8 @@
-package io.github.willqi.pizzaserver.server.plugin.events.player;
+package io.github.willqi.pizzaserver.server.event.type.player;
 
 import io.github.willqi.pizzaserver.server.player.Player;
 
-public class PreLoginEvent extends PlayerEvent {
+public class PreLoginEvent extends PlayerEvent.Cancellable {
 
     public PreLoginEvent(Player player) {
         super(player);
